airflow/requirements.txt
apache-airflow==2.7.3
psycopg2-binary
asyncpg
python-dotenv
requests
redis
pendulum
pandas
xlsxwriter